+++ test/dates.test.ts
@@ -1,7 +1,7 @@
 import { partition } from 'itertools';
 import { describe, expect, test } from 'vitest';
 
-import { date, datelike, iso8601 } from '~/dates';
+import { date, datelike, dateString, iso8601 } from '~/dates';
 
 import { INPUTS } from './_fixtures';
 
@@ -24,6 +24,42 @@ describe('pure dates', () => {
     for (const value of not_okay) {
       expect(() => decoder.verify(value)).toThrow();
     }
+  });
+});
+
+describe('dateString', () => {
+  const decoder = dateString;
+
+  test('invalid', () => {
+    // None of the values in INPUTS are valid ISO8601 strings
+    const not_okay = INPUTS;
+    for (const value of not_okay) {
+      expect(() => decoder.verify(value)).toThrow();
+    }
+  });
+
+  test('decodes ISO dates', () => {
+    expect(decoder.verify('2020-06-22T10:57:33Z')).toEqual('2020-06-22T10:57:33Z');
+    expect(decoder.verify('2020-06-22T10:57:33+02:00')).toEqual(
+      '2020-06-22T10:57:33+02:00',
+    );
+
+    // Note: Feb 30 does not exist, but a new Date() constructor would "fix"
+    // that, so we do consider it valid
+    expect(decoder.verify('2020-02-30T10:57:33+02:00')).toEqual(
+      '2020-02-30T10:57:33+02:00',
+    );
+  });
+
+  test('rejects invalid dates', () => {
+    // Syntactically invalid
+    expect(() => decoder.verify('03/04/2000')).toThrow();
+    expect(() => decoder.verify('2020-06-22T10:57:33')).toThrow();
+    expect(() => decoder.verify('2020-06-22')).toThrow();
+
+    // Semantically invalid (these dates don't exist)
+    expect(() => decoder.verify('2020-03-32T10:57:33Z')).toThrow();
+    expect(() => decoder.verify('0099-16-48T10:57:33Z')).toThrow();
   });
 });